+++ container-engine-lib/docs/changelog.md
@@ -1,4 +1,8 @@
 # TBD
+### Changes
+* Migrated the `engine functions` from `docker_kurtosis_backend` package to the `engine_functions` package
+* Removed duplicated shared functions that where present in the both the `docker_kurtosis_backend` package and the `shared_helpers` package
+* Added the `consts` package to group the general constants used in the Docker Kurtosis backend
 
 # 0.34.0